Change groupid to io.opentracing.contrib
