+++ README.md
@@ -1,13 +1,7 @@
-# OpenTracing Java Integrations
+# OpenTracing JAX-RS 2.0 Instrumentation
 
 Warning: This library is still a work in progress!
 
 [![Travis](https://travis-ci.org/pavolloffay/opentracing-java-integrations.svg?branch=master)](https://travis-ci.org/pavolloffay/opentracing-java-integrations)
 
-
-This repository will provide instrumentation of several Java frameworks using OpenTracing API.
-Consumers of this libraries will provide OpenTracing API implementation, therefore this 
-instrumentation does not depend on any tracer implementation.
-
-## Supported Frameworks
-* JAX-RS
+This repository provides instrumentation for JAX-RS API.
