+++ examples/dropwizard/pom.xml
@@ -4,13 +4,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>sk.loffay.opentracing</groupId>
-    <artifactId>opentracing-integration-parent</artifactId>
+    <groupId>io.opentracing.contrib</groupId>
+    <artifactId>opentracing-jaxrs-parent</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <relativePath>../../pom.xml</relativePath>
   </parent>
 
-  <artifactId>opentracing-integration-example-dropwizard</artifactId>
+  <artifactId>opentracing-jaxrs-example-dropwizard</artifactId>
 
   <properties>
     <version.org.hawkular.apm>0.12.0.Final</version.org.hawkular.apm>
@@ -43,7 +43,7 @@
     <!-- OpenTracing -->
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>opentracing-integration-jax-rs</artifactId>
+      <artifactId>opentracing-jaxrs</artifactId>
     </dependency>
 
     <!-- Your preferred OpenTracing impl -->
@@ -88,7 +88,7 @@
               <transformers>
                 <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <mainClass>sk.loffay.opentracing.example.dropwizard.App</mainClass>
+                  <mainClass>io.opentracing.contrib.jaxrs.example.dropwizard.App</mainClass>
                 </transformer>
               </transformers>
               <finalName>${project.artifactId}</finalName>
