+++ examples/dropwizard/src/main/java/io/opentracing/contrib/jaxrs/example/dropwizard/App.java
@@ -1,0 +1,75 @@
+/*
+ * Copyright 2015-2016 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.opentracing.contrib.jaxrs.example.dropwizard;
+
+import javax.ws.rs.client.Client;
+
+import org.glassfish.hk2.utilities.binding.AbstractBinder;
+import org.glassfish.jersey.client.JerseyClientBuilder;
+import org.hawkular.apm.client.opentracing.APMTracer;
+
+import io.dropwizard.Application;
+import io.dropwizard.configuration.EnvironmentVariableSubstitutor;
+import io.dropwizard.configuration.SubstitutingSourceProvider;
+import io.dropwizard.setup.Bootstrap;
+import io.dropwizard.setup.Environment;
+import io.opentracing.Tracer;
+import io.opentracing.contrib.jaxrs.client.ClientTracingFeature;
+import io.opentracing.contrib.jaxrs.server.ServerTracingDynamicFeature;
+
+/**
+ * @author Pavol Loffay
+ */
+public class App extends Application<AppConfiguration> {
+
+    public static void main(String[] args) throws Exception {
+        new App().run(args);
+    }
+
+    @Override
+    public String getName() {
+        return "dropwizard";
+    }
+
+    @Override
+    public void initialize(Bootstrap<AppConfiguration> bootstrap) {
+        bootstrap.setConfigurationSourceProvider(
+                new SubstitutingSourceProvider(bootstrap.getConfigurationSourceProvider(),
+                        new EnvironmentVariableSubstitutor()));
+    }
+
+    @Override
+    public void run(AppConfiguration configuration, Environment environment) throws Exception {
+        Tracer tracer = new APMTracer();
+        Client client = ClientTracingFeature.Builder
+                .traceAll(tracer, new JerseyClientBuilder().build())
+                .build();
+
+        environment.jersey().register(ServerTracingDynamicFeature.Builder.traceAll(tracer).build());
+        environment.jersey().register(new AbstractBinder() {
+            @Override
+            protected void configure() {
+//                TODO remove
+                bind(APMTracer.class).to(Tracer.class);
+            }
+        });
+
+        // Register resources
+        environment.jersey().register(new HelloHandler(tracer, client));
+    }
+}
