+++ examples/spring-boot/src/main/java/io/opentracing/contrib/jaxrs/example/spring/boot/App.java
@@ -1,0 +1,16 @@
+package io.opentracing.contrib.jaxrs.example.spring.boot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContext;
+
+/**
+ * @author Pavol Loffay
+ */
+@SpringBootApplication
+public class App {
+    public static void main(String[] args) {
+        ApplicationContext ctx = SpringApplication.run(App.class, args);
+    }
+
+}
