+++ examples/spring-boot/src/main/java/io/opentracing/contrib/jaxrs/example/spring/boot/Configuration.java
@@ -1,0 +1,18 @@
+package io.opentracing.contrib.jaxrs.example.spring.boot;
+
+import org.hawkular.apm.client.opentracing.APMTracer;
+import org.springframework.context.annotation.Bean;
+
+import io.opentracing.Tracer;
+
+/**
+ * @author Pavol Loffay
+ */
+@org.springframework.context.annotation.Configuration
+public class Configuration {
+
+    @Bean
+    public Tracer tracer() {
+        return new APMTracer();
+    }
+}
