--- examples/spring-boot/src/main/java/sk/loffay/opentracing/example/spring/boot/Configuration.java
@@ -1,18 +1,0 @@
-package sk.loffay.opentracing.example.spring.boot;
-
-import org.hawkular.apm.client.opentracing.APMTracer;
-import org.springframework.context.annotation.Bean;
-
-import io.opentracing.Tracer;
-
-/**
- * @author Pavol Loffay
- */
-@org.springframework.context.annotation.Configuration
-public class Configuration {
-
-    @Bean
-    public Tracer tracer() {
-        return new APMTracer();
-    }
-}
