--- jax-rs/src/main/java/sk/loffay/opentracing/jax/rs/server/CurrentSpan.java
@@ -1,20 +1,0 @@
-package sk.loffay.opentracing.jax.rs.server;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.ws.rs.core.Context;
-
-import io.opentracing.Span;
-
-
-/**
- * @author Pavol Loffay
- */
-public class CurrentSpan {
-
-    @Context
-    private HttpServletRequest request;
-
-    public Span injectedSpan() {
-        return (Span) request.getAttribute(SpanServerRequestFilter.SPAN_PROP_ID);
-    }
-}
