--- jax-rs/src/main/java/sk/loffay/opentracing/jax/rs/server/Traced.java
@@ -1,17 +1,0 @@
-package sk.loffay.opentracing.jax.rs.server;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Annotation to explicitly say that tracing is enabled.
- *
- * @author Pavol Loffay
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD, ElementType.TYPE})
-public @interface Traced {
-    String operationName();
-}
