+++ jaxrs/src/main/java/io/opentracing/contrib/jaxrs/server/CurrentSpan.java
@@ -1,0 +1,20 @@
+package io.opentracing.contrib.jaxrs.server;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.core.Context;
+
+import io.opentracing.Span;
+
+
+/**
+ * @author Pavol Loffay
+ */
+public class CurrentSpan {
+
+    @Context
+    private HttpServletRequest request;
+
+    public Span injectedSpan() {
+        return (Span) request.getAttribute(SpanServerRequestFilter.SPAN_PROP_ID);
+    }
+}
