+++ jaxrs/src/main/java/io/opentracing/contrib/jaxrs/server/Traced.java
@@ -1,0 +1,17 @@
+package io.opentracing.contrib.jaxrs.server;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Annotation to explicitly say that tracing is enabled.
+ *
+ * @author Pavol Loffay
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD, ElementType.TYPE})
+public @interface Traced {
+    String operationName();
+}
