+++ jaxrs/src/main/resources/META-INF/beans.xml
